requirements.txt
discord.py==1.7.3
colorama==0.4.6
aioconsole==0.4.0
